+++ src/main/java/com/wenxin2/marioverse/datagen/ItemModelGen.java
@@ -124,6 +124,7 @@ public class ItemModelGen extends ItemModelProvider {
 
         this.handheldItem(BlockRegistry.CLASSIC_CHECKPOINT_FLAG.asItem());
         this.handheldItem(BlockRegistry.CLASSIC_GOAL_POLE.asItem());
+        this.handheldItem(ItemRegistry.CREATIVE_WRENCH.get());
         this.handheldItem(ItemRegistry.WRENCH.get());
         this.handheldItem(ItemRegistry.WARP_DISRUPTOR.get());