+++ changelog.md
@@ -1,3 +1,5 @@
 ## Changelog for Marioverse 1.7.2
+- Add Creative Wrench
+  - Links any entities together 
 - Add chipped mushrooms to bouncy tag
 - Fix Dango Blossom missing from Building Blocks Creative Tab